# Refactor Emval implementation to avoid JS heap allocations.