--- a/test/code_size/embind_val_wasm.json
+++ b/test/code_size/embind_val_wasm.json
@@ -1,10 +1,10 @@
 {
   "a.html": 673,
   "a.html.gz": 431,
-  "a.js": 7155,
-  "a.js.gz": 3022,
+  "a.js": 7160,
+  "a.js.gz": 3027,
   "a.wasm": 11433,
   "a.wasm.gz": 5725,
-  "total": 19261,
-  "total_gz": 9178
+  "total": 19266,
+  "total_gz": 9183
 }