--- a/test/code_size/embind_val_wasm.json
+++ b/test/code_size/embind_val_wasm.json
@@ -1,10 +1,10 @@
 {
   "a.html": 673,
   "a.html.gz": 431,
-  "a.js": 7098,
-  "a.js.gz": 3001,
+  "a.js": 7089,
+  "a.js.gz": 3002,
   "a.wasm": 11433,
   "a.wasm.gz": 5725,
-  "total": 19204,
-  "total_gz": 9157
+  "total": 19195,
+  "total_gz": 9158
 }