Refactors the JS implementation of emscripten::val to use a single array to hold both the JavaScript values and the refcount, rather than using the array from a HandleAllocator with JS objects holding pairs of value and refcount.

Motivation is to reduce overhead for allocating such objects and GC pressure from short-lived objects. This design continues to use a single array that never shrinks to hold the JS values an refcounts, and a freelist as a stack of available handles to reuse array entries. Once in a steady state, there should be approximately no allocations required to add and remove entries.

This changes Emval to not reuse the HandleAllocator, so some logic from that class is duplicated, with revision for the array to hold two entries per handle.

In my target application, this was the only use of HandleAllocator, and I saw a decrease in JS code size as a result (both before and after closure compilation). Even when I left in the dependency on HandleAllocator to keep that code in the binary to approximate impact of projects that still use HandleAllocator, I observe a smaller code size with this change.

I do not have a benchmark to demonstrate the improvement, but believe on the surface avoiding heap allocations is obviously better, especially if there isn't code size growth as a result.

/cc @brendandahl @sbc100