@RReverser 2024-01-29
Hm it would be good to see some benchmarks before/after this change. JS engines generally optimise frequent creation / deletion of small objects very well (because it's a very common usage), and they have extra optimisations for arrays of monomorphic shape, so I'd expect GC pressure not to come into play here at all, unlike in interpreted languages.

Given that this somewhat lowers readability of the code and makes array non-monomorphic, I think some simple benchmarks showing a clear performance win would be warranted before merging.