@RReverser 2024-01-29
This also reminds me that I should find time (if only) and finish https://github.com/emscripten-core/emscripten/pull/20447 to move reference counting to C++ entirely...