@mrolig5267319 2024-01-30
@RReverser Thanks for the reference to your other [draft pull request](https://github.com/emscripten-core/emscripten/pull/20447) with a benchmark.

I've run that benchmark and I observe a ~20% improvement to Allocation cost:

Before:

```
Allocation cost: 92 ns
Copying cost: 61 ns
```

After:

```
Allocation cost: 76 ns
Copying cost: 59 ns
```