@mrolig5267319 2024-01-30
Fixed test_embind_val ... the reason we special case the increment/decrement for reserved values is that C++ will use them without incrementing the refcount.  Had to add that back to the JS.